AI content generation works well. Problems with pptx parsing/modification. Results in corupt pptx file. See logs
