+++ backend/app/api/endpoints/slides.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import Session
 from typing import List
 from app.schemas import schemas
 from app.database import get_db
-from app.utils.pptx import process_powerpoint_repository
+from app.utils.pptx_parsing import process_powerpoint_repository
 from app.models.models import PresentationMetadata, SlideMetadata
 from app.utils.openai import get_embedding
 import json
@@ -13,11 +13,11 @@ router = APIRouter()
 @router.post("/slides/repository/upload")
 async def upload_slide_repository(
     file: UploadFile = File(...),
-    title: str = Form(...),
+    title: str | None = Form(None),
     db: Session = Depends(get_db)
 ):
-    """Upload a PowerPoint file to create/update the slide repository"""
-    # Process the PowerPoint file
+    """Upload a PowerPoint file to create/update the slide repository and presentation metadata as well as the slide metadata, embedding of metadata, and content schema"""
+
     storage_path, slide_metadata_objects = await process_powerpoint_repository(
         file.file, 
         db, 
@@ -25,7 +25,8 @@ async def upload_slide_repository(
     )
 
     # Create presentation metadata
-    presentation = PresentationMetadata(storage_path=storage_path, title=title)
+    presentation_title = title or "Untitled Slide Repository"
+    presentation = PresentationMetadata(storage_path=storage_path, title=presentation_title)
     db.add(presentation)
     db.flush()
     
@@ -42,14 +43,7 @@ async def upload_slide_repository(
         "presentation_id": presentation.id
     }
 
-@router.post("/slides/repository/sync")
-async def sync_slide_repository(
-    file_id: str,
-    db: Session = Depends(get_db)
-):
-    """Sync with the PowerPoint file in Microsoft 365"""
-    # result = await utils.slide_repository.sync_repository(file_id, db)
-    # return result
+
 
 @router.get("/slides/metadata/{presentation_id}", response_model=List[schemas.SlideMetadata])
 async def get_slides(
@@ -59,6 +53,8 @@ async def get_slides(
     """Get all slide metadata for a presentation"""
     presentation = db.query(PresentationMetadata).filter(PresentationMetadata.id == presentation_id).first()
     return presentation.slides
+
+
 
 @router.put("/slides/metadata/{id}")
 async def update_slide_metadata(
@@ -84,3 +80,14 @@ async def update_slide_metadata(
     slide.embedding = get_embedding(stringified_metadata)
     db.commit()
     return slide
+
+
+
+# @router.post("/slides/repository/sync")
+# async def sync_slide_repository(
+#     file_id: str,
+#     db: Session = Depends(get_db)
+# ):
+#     """Sync with the PowerPoint file in Microsoft 365"""
+    # result = await utils.slide_repository.sync_repository(file_id, db)
+    # return result
