+++ backend/app/config.py
@@ -8,7 +8,7 @@ load_dotenv()
 class Settings(BaseSettings):
     API_PREFIX: str = "/api/v1"
     DEBUG: bool = False
-    PROJECT_NAME: str = "FastAPI with pgvector"
+    PROJECT_NAME: str = "AI PowerPoint Presentation Builder"
     DATABASE_URI: str = os.getenv("DATABASE_URI", "postgresql://postgres:postgres@db:5432/myapp")
     SECRET_KEY: str = os.getenv("SECRET_KEY", "default-secret-key")
     OPENAI_API_KEY: str = ""
@@ -22,7 +22,6 @@ class Settings(BaseSettings):
 
     @field_validator("OPENAI_API_KEY", mode="before")
     def validate_openai_api_key(cls, v):
-        # This allows the API key to be loaded from environment variables
         return os.getenv("OPENAI_API_KEY", v)
 
 settings = Settings()
