+++ backend/app/models/models.py
@@ -22,7 +22,7 @@ class SlideMetadata(Base):
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
 
-    content_schema = Column(JSON)  # Defines structure for content replacement
+    content_mapping = Column(JSON)  # Defines structure for content replacement
 
     presentation = relationship("PresentationMetadata", back_populates="slides")
