+++ backend/app/schemas/schemas.py
@@ -1,6 +1,6 @@
 from typing import List, Optional, Dict, Any
 from datetime import datetime
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 class SlideMetadataBase(BaseModel):
     title: Optional[str] = None
@@ -10,11 +10,11 @@ class SlideMetadataBase(BaseModel):
     tags: Optional[List[str]] = None
     audience: Optional[str] = None
     sales_stage: Optional[str] = None
-    content_schema: Optional[Dict[str, Any]] = None
+    content_mapping: Optional[Dict[str, Any]] = None
 
 class SlideMetadataUpdate(SlideMetadataBase):
     class Config:
-        exclude = {'content_schema'}
+        exclude = {'content_mapping'}
 
 class SlideMetadata(SlideMetadataBase):
     id: int
@@ -76,3 +76,31 @@ class PresentationInput(BaseModel):
     preferred_slide_types: Optional[List[str]] = None
     tone: Optional[str] = None
     additional_context: Optional[str] = None
+
+# For OpenAI response parsing
+class SlideOutlineBase(BaseModel):
+    slide_number: int
+    section: str
+    description: str
+    keywords: List[str] | None = None
+
+class PresentationOutlineResponse(BaseModel):
+    slides: List[SlideOutlineBase]
+
+# For API validation
+class SlideOutline(SlideOutlineBase):
+    slide_number: int = Field(description="The sequential position in the presentation", ge=1)
+    section: str = Field(description="A general category or purpose for this slide")
+    description: str = Field(description="A descriptive summary of what this slide should communicate")
+    keywords: List[str] | None = Field(default=None, description="Keywords for semantic matching with template repository")
+
+# models for slide content generation
+class SlideContentBase(BaseModel):
+    title: str | None = None
+    subtitle: str | None = None
+    body: str | None = None
+    notes: str | None = None
+    #TODO: Add any other common fields that might be in content_mapping
+
+class SlideContentResponse(BaseModel):
+    content: SlideContentBase
